refactor(x86_64): move firecracker and multiboot into modules
